requirements.txt
langchain-groq
langchain-huggingface
langchain
langchain-community
python-dotenv
langchain-chroma>=0.1.2
pypdf